pyproject.toml
[tool.poetry]
name = "dataflow-example"
version = "0.1.0"
description = ""
authors = ["yes"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
apache-airflow = "2.7.3"
boto3 = "1.34.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
